# feat(testing): Comprehensive Performance Testing Suite - Issue #176