--- a/src-tauri/tests/e2e_ollama_tests.rs
+++ b/src-tauri/tests/e2e_ollama_tests.rs
@@ -151,6 +151,12 @@ mod e2e_tests {
             max_retries: 3,
             initial_retry_delay_ms: 50, // Short delays for faster test
             max_retry_delay_ms: 200,
+            enable_http2_pipelining: true,
+            max_concurrent_requests: 6,
+            keep_alive_seconds: 120,
+            enable_request_batching: true,
+            max_batch_size: 10,
+            batch_timeout_ms: 50,
         };
         let client = OllamaClient::with_config(config);
         
@@ -378,6 +384,12 @@ mod e2e_tests {
             max_retries: 2,
             initial_retry_delay_ms: 100,
             max_retry_delay_ms: 1000,
+            enable_http2_pipelining: true,
+            max_concurrent_requests: 6,
+            keep_alive_seconds: 120,
+            enable_request_batching: true,
+            max_batch_size: 10,
+            batch_timeout_ms: 50,
         };
         let client = OllamaClient::with_config(config);