--- a/src-tauri/tests/frontend_integration_simplified.rs
+++ b/src-tauri/tests/frontend_integration_simplified.rs
@@ -52,6 +52,12 @@ mod client_integration_tests {
                 max_retries: 3,
                 initial_retry_delay_ms: 1000,
                 max_retry_delay_ms: 30000,
+                enable_http2_pipelining: true,
+                max_concurrent_requests: 6,
+                keep_alive_seconds: 120,
+                enable_request_batching: true,
+                max_batch_size: 10,
+                batch_timeout_ms: 50,
             };
             
             client.update_config(config).await;
@@ -68,6 +74,12 @@ mod client_integration_tests {
             max_retries: 0,
             initial_retry_delay_ms: 100,
             max_retry_delay_ms: 1000,
+            enable_http2_pipelining: true,
+            max_concurrent_requests: 6,
+            keep_alive_seconds: 120,
+            enable_request_batching: true,
+            max_batch_size: 10,
+            batch_timeout_ms: 50,
         };
         
         let client = OllamaClient::with_config(config);