… tests/comprehensive/README.md
… tests/comprehensive/performance-validation.test.js
… tests/comprehensive/regression-detection.test.js
… tests/comprehensive/stress-testing.test.js